brdecode: add support for compression format 1032
This format may be used by the Windows driver when printing in 1200 dpi
or HQ1200 mode. It is very similar to the existing 1030 format, but
operates on 16-bit data values instead of 8-bit data values.

The edit counter at the start of each line is also extended to use two
bytes, but it is not clear how exactly it is constructed. For now, only
the second byte is actually used. An exception is generated if the first
byte has an unexpected value, so this can be revisited when someone
encounters a print file triggering it.